Renamed StrategyEditorPage to StrategiesPage
