+++ src/pages/Strategies/index.js
@@ -1,0 +1,3 @@
+import StrategiesPage from './Strategies.container'
+
+export default StrategiesPage
