+++ src/pages/Strategies/mock_data.js
