+++ src/pages/Strategies/style.scss
