--- src/pages/StrategyEditor/index.js
@@ -1,3 +1,0 @@
-import StrategyEditor from './StrategyEditor.container'
-
-export default StrategyEditor
